# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Map of commission type to rate percentage (e.g. FYC: 15 means 15%)
 */
export type AppSettingsCommissionRates = { [key: string]: number };

export interface AppSettings {
  /** 0=Sun 1=Mon 2=Tue 3=Wed 4=Thu 5=Fri 6=Sat */
  reportDayOfWeek: number;
  /** Hour 0-23 */
  reportHour: number;
  /** Minute 0-59 */
  reportMinute: number;
  recipients: string[];
  /** Map of commission type to rate percentage (e.g. FYC: 15 means 15%) */
  commissionRates: AppSettingsCommissionRates;
}

export type UpdateSettingsBodyCommissionRates = { [key: string]: number };

export interface UpdateSettingsBody {
  reportDayOfWeek?: number;
  reportHour?: number;
  reportMinute?: number;
  recipients?: string[];
  commissionRates?: UpdateSettingsBodyCommissionRates;
}

export interface SaleEntry {
  id: number;
  clientName: string;
  owningAgent: string;
  salesType: string;
  soldDate: string;
  commissionType: string;
  /** @nullable */
  annualPremium: number | null;
  /** @nullable */
  estimatedCommission: number | null;
  /** @nullable */
  notes: string | null;
  weekStart: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSaleBody {
  clientName: string;
  owningAgent: string;
  salesType: string;
  soldDate: string;
  commissionType: string;
  /** @nullable */
  annualPremium?: number | null;
  /** @nullable */
  estimatedCommission?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateSaleBody {
  clientName?: string;
  owningAgent?: string;
  salesType?: string;
  soldDate?: string;
  commissionType?: string;
  /** @nullable */
  annualPremium?: number | null;
  /** @nullable */
  estimatedCommission?: number | null;
  /** @nullable */
  notes?: string | null;
}

export type WeekSummaryByOwningAgentItem = {
  agent: string;
  count: number;
};

export type WeekSummaryBySalesTypeItem = {
  salesType: string;
  count: number;
};

export interface WeekSummary {
  totalSales: number;
  totalEstimatedCommission: number;
  byOwningAgent: WeekSummaryByOwningAgentItem[];
  bySalesType: WeekSummaryBySalesTypeItem[];
  weekStart: string;
  weekEnd: string;
}

export interface WeeklyReport {
  id: number;
  weekStart: string;
  weekEnd: string;
  sentAt: string;
  totalSales: number;
  totalEstimatedCommission: number;
  recipients: string;
}

export interface SendReportResponse {
  message: string;
  reportId: number;
}

export type ListSalesParams = {
  /**
   * ISO date string for the start of the week filter
   */
  weekStart?: string;
};