# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: sales
    description: Sales entry operations
  - name: reports
    description: Weekly report operations
  - name: settings
    description: App configuration settings
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List all sales entries
      parameters:
        - name: weekStart
          in: query
          required: false
          schema:
            type: string
          description: ISO date string for the start of the week filter
      responses:
        "200":
          description: List of sales entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SaleEntry"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a new sale entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleBody"
      responses:
        "201":
          description: Created sale entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleEntry"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get a single sale entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSale
      tags: [sales]
      summary: Update a sale entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSaleBody"
      responses:
        "200":
          description: Updated sale entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete a sale entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sales/summary/current-week:
    get:
      operationId: getCurrentWeekSummary
      tags: [sales]
      summary: Get summary stats for the current week
      responses:
        "200":
          description: Weekly summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeekSummary"

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all weekly reports
      responses:
        "200":
          description: List of sent weekly reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeeklyReport"

  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get current app settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update app settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reports/send:
    post:
      operationId: sendReport
      tags: [reports]
      summary: Manually trigger sending the weekly report now
      responses:
        "200":
          description: Report sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendReportResponse"
        "400":
          description: Error sending report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    AppSettings:
      type: object
      properties:
        reportDayOfWeek:
          type: integer
          description: "0=Sun 1=Mon 2=Tue 3=Wed 4=Thu 5=Fri 6=Sat"
        reportHour:
          type: integer
          description: "Hour 0-23"
        reportMinute:
          type: integer
          description: "Minute 0-59"
        recipients:
          type: array
          items:
            type: string
        commissionRates:
          type: object
          additionalProperties:
            type: number
          description: "Map of commission type to rate percentage (e.g. FYC: 15 means 15%)"
      required:
        - reportDayOfWeek
        - reportHour
        - reportMinute
        - recipients
        - commissionRates

    UpdateSettingsBody:
      type: object
      properties:
        reportDayOfWeek:
          type: integer
        reportHour:
          type: integer
        reportMinute:
          type: integer
        recipients:
          type: array
          items:
            type: string
        commissionRates:
          type: object
          additionalProperties:
            type: number

    SaleEntry:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        owningAgent:
          type: string
        salesType:
          type: string
        soldDate:
          type: string
        commissionType:
          type: string
        annualPremium:
          type: ["number", "null"]
        estimatedCommission:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        weekStart:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientName
        - owningAgent
        - salesType
        - soldDate
        - commissionType
        - annualPremium
        - estimatedCommission
        - notes
        - weekStart
        - createdAt
        - updatedAt

    CreateSaleBody:
      type: object
      properties:
        clientName:
          type: string
        owningAgent:
          type: string
        salesType:
          type: string
        soldDate:
          type: string
        commissionType:
          type: string
        annualPremium:
          type: ["number", "null"]
        estimatedCommission:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
      required:
        - clientName
        - owningAgent
        - salesType
        - soldDate
        - commissionType

    UpdateSaleBody:
      type: object
      properties:
        clientName:
          type: string
        owningAgent:
          type: string
        salesType:
          type: string
        soldDate:
          type: string
        commissionType:
          type: string
        annualPremium:
          type: ["number", "null"]
        estimatedCommission:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    WeekSummary:
      type: object
      properties:
        totalSales:
          type: integer
        totalEstimatedCommission:
          type: number
        byOwningAgent:
          type: array
          items:
            type: object
            properties:
              agent:
                type: string
              count:
                type: integer
            required:
              - agent
              - count
        bySalesType:
          type: array
          items:
            type: object
            properties:
              salesType:
                type: string
              count:
                type: integer
            required:
              - salesType
              - count
        weekStart:
          type: string
        weekEnd:
          type: string
      required:
        - totalSales
        - totalEstimatedCommission
        - byOwningAgent
        - bySalesType
        - weekStart
        - weekEnd

    WeeklyReport:
      type: object
      properties:
        id:
          type: integer
        weekStart:
          type: string
        weekEnd:
          type: string
        sentAt:
          type: string
        totalSales:
          type: integer
        totalEstimatedCommission:
          type: number
        recipients:
          type: string
      required:
        - id
        - weekStart
        - weekEnd
        - sentAt
        - totalSales
        - totalEstimatedCommission
        - recipients

    SendReportResponse:
      type: object
      properties:
        message:
          type: string
        reportId:
          type: integer
      required:
        - message
        - reportId