# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all sales entries
 */
export const ListSalesQueryParams = zod.object({
  weekStart: zod.coerce
    .string()
    .optional()
    .describe("ISO date string for the start of the week filter"),
});

export const ListSalesResponseItem = zod.object({
  id: zod.number(),
  clientName: zod.string(),
  owningAgent: zod.string(),
  salesType: zod.string(),
  soldDate: zod.string(),
  commissionType: zod.string(),
  annualPremium: zod.number().nullable(),
  estimatedCommission: zod.number().nullable(),
  notes: zod.string().nullable(),
  weekStart: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListSalesResponse = zod.array(ListSalesResponseItem);

/**
 * @summary Create a new sale entry
 */
export const CreateSaleBody = zod.object({
  clientName: zod.string(),
  owningAgent: zod.string(),
  salesType: zod.string(),
  soldDate: zod.string(),
  commissionType: zod.string(),
  annualPremium: zod.number().nullish(),
  estimatedCommission: zod.number().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a single sale entry
 */
export const GetSaleParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSaleResponse = zod.object({
  id: zod.number(),
  clientName: zod.string(),
  owningAgent: zod.string(),
  salesType: zod.string(),
  soldDate: zod.string(),
  commissionType: zod.string(),
  annualPremium: zod.number().nullable(),
  estimatedCommission: zod.number().nullable(),
  notes: zod.string().nullable(),
  weekStart: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a sale entry
 */
export const UpdateSaleParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSaleBody = zod.object({
  clientName: zod.string().optional(),
  owningAgent: zod.string().optional(),
  salesType: zod.string().optional(),
  soldDate: zod.string().optional(),
  commissionType: zod.string().optional(),
  annualPremium: zod.number().nullish(),
  estimatedCommission: zod.number().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateSaleResponse = zod.object({
  id: zod.number(),
  clientName: zod.string(),
  owningAgent: zod.string(),
  salesType: zod.string(),
  soldDate: zod.string(),
  commissionType: zod.string(),
  annualPremium: zod.number().nullable(),
  estimatedCommission: zod.number().nullable(),
  notes: zod.string().nullable(),
  weekStart: zod.string(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a sale entry
 */
export const DeleteSaleParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get summary stats for the current week
 */
export const GetCurrentWeekSummaryResponse = zod.object({
  totalSales: zod.number(),
  totalEstimatedCommission: zod.number(),
  byOwningAgent: zod.array(
    zod.object({
      agent: zod.string(),
      count: zod.number(),
    }),
  ),
  bySalesType: zod.array(
    zod.object({
      salesType: zod.string(),
      count: zod.number(),
    }),
  ),
  weekStart: zod.string(),
  weekEnd: zod.string(),
});

/**
 * @summary List all weekly reports
 */
export const ListReportsResponseItem = zod.object({
  id: zod.number(),
  weekStart: zod.string(),
  weekEnd: zod.string(),
  sentAt: zod.string(),
  totalSales: zod.number(),
  totalEstimatedCommission: zod.number(),
  recipients: zod.string(),
});
export const ListReportsResponse = zod.array(ListReportsResponseItem);

/**
 * @summary Get current app settings
 */
export const GetSettingsResponse = zod.object({
  reportDayOfWeek: zod
    .number()
    .describe("0=Sun 1=Mon 2=Tue 3=Wed 4=Thu 5=Fri 6=Sat"),
  reportHour: zod.number().describe("Hour 0-23"),
  reportMinute: zod.number().describe("Minute 0-59"),
  recipients: zod.array(zod.string()),
  commissionRates: zod
    .record(zod.string(), zod.number())
    .describe(
      "Map of commission type to rate percentage (e.g. FYC: 15 means 15%)",
    ),
});

/**
 * @summary Update app settings
 */
export const UpdateSettingsBody = zod.object({
  reportDayOfWeek: zod.number().optional(),
  reportHour: zod.number().optional(),
  reportMinute: zod.number().optional(),
  recipients: zod.array(zod.string()).optional(),
  commissionRates: zod.record(zod.string(), zod.number()).optional(),
});

export const UpdateSettingsResponse = zod.object({
  reportDayOfWeek: zod
    .number()
    .describe("0=Sun 1=Mon 2=Tue 3=Wed 4=Thu 5=Fri 6=Sat"),
  reportHour: zod.number().describe("Hour 0-23"),
  reportMinute: zod.number().describe("Minute 0-59"),
  recipients: zod.array(zod.string()),
  commissionRates: zod
    .record(zod.string(), zod.number())
    .describe(
      "Map of commission type to rate percentage (e.g. FYC: 15 means 15%)",
    ),
});

/**
 * @summary Manually trigger sending the weekly report now
 */
export const SendReportResponse = zod.object({
  message: zod.string(),
  reportId: zod.number(),
});